Sync bootstrap glyphicons from bower_components folder on every build. (Improvement to fix #24)
If there is any change to the glyphicons files we would have to update them manually. This fix allows a user to simply update the bootstrap package via bower and everything else just works. This could also be used for syncing various other static files like the fonts of font-awesome.

Signed-off-by: David Gstir <david@sigma-star.at>